Restructure, pull out src and add transform data folder
