+++ 1-raw-data/mailinglists/pipermail/pipermail_data_test.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/google/project-OCEAN/src/1-raw-data/gcs"
+	"github.com/google/project-OCEAN/1-raw-data/gcs"
 )
 
 type fakeStorageConnection struct {
