+++ sql/Ex_Bigquery_SQL_Info.md
